backend/requirements/base.txt
pytz==2018.5  # https://github.com/stub42/pytz
python-slugify==1.2.6  # https://github.com/un33k/python-slugify
Pillow==5.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.3.0  # https://github.com/hynek/argon2_cffi
redis >= 3.2.0  # https://github.com/antirez/redis
celery==4.2.1  # pyup: <5.0  # https://github.com/celery/celery
flower==0.9.3  # https://github.com/mher/flower
anyjson
Faker==0.9.2
PyYAML==3.13
python-dateutil==2.8.0

# Django
# ------------------------------------------------------------------------------
Django==2.1
django-cors-headers==2.4.0

# Django
# ------------------------------------------------------------------------------
Django==2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.37.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-threadedcomments==1.2 # https://github.com/HonzaKral/django-threadedcomments
django-notifications-hq==1.5.0 # https://github.com/django-notifications/django-notifications
django-admin-rangefilter==0.3.12 # https://github.com/silentsokolov/django-admin-rangefilter#django-admin-rangefilter
django-import-export==1.2.0 # https://github.com/django-import-export/django-import-export
django-constance[database]==2.4.0

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-filter==2.0.0
django-rest-auth[with_social]==0.9.3
djangorestframework_jwt==1.11.0
djangorestframework-csv==2.1.0

markdown==2.6.11

# ImageField
django-versatileimagefield==1.9

# Stripe
stripe==2.15.0

python-magic==0.4.15